# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: OmniRepo API
  version: "1.0.0"
  description: |
    REST API for OmniRepo — self-hosted artifact repository server.
    Serves OCI/Docker, RPM, APT/Debian, PyPI, Helm, RAW, S3, and Git
    from a single container on one HTTP/HTTPS port.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ApiErrorClass:
      type: string
      enum: [validation, permission, transient, operator_action_required]
      description: Stable client-facing error classification — drives UI icon/copy/CTA.
    ApiErrorEnvelope:
      type: object
      required: [code, message, class]
      properties:
        code:
          type: string
          pattern: "^[a-z][a-z0-9_]*\\.[a-z][a-z0-9_]*$"
          maxLength: 80
          description: Stable machine identifier (e.g. "repo.not_found")
        message:
          type: string
          maxLength: 280
          description: Human-friendly single-sentence message (no internal paths)
        hint:
          type: string
          maxLength: 280
          description: Optional single-sentence remediation guidance
        class:
          $ref: '#/components/schemas/ApiErrorClass'
        incident_id:
          type: string
          maxLength: 64
          description: Server-generated correlation ID (UUID v7) — matches slog request_id and the X-Incident-Id response header
        details:
          type: object
          additionalProperties: true
          properties:
            field:
              type: string
              description: Dot-path of the offending form field (validation class)
            fields:
              type: object
              additionalProperties:
                type: string
              description: Map of field-path to error-code for multi-field validation failures
            retry_after_ms:
              type: integer
              minimum: 0
              description: Suggested retry delay in ms (transient class)
            operator_route:
              type: string
              description: UI path (e.g. "/admin/trivy") for operator_action_required deep-link
            operator_label:
              type: string
              description: CTA label for the operator deep-link (e.g. "Go to Admin → Trivy")
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        email:
          type: string
        is_super_admin:
          type: boolean
        must_change_password:
          type: boolean
        avatar_seed:
          type: string
        created_at:
          type: string
          format: date-time
...
    UserCreate:
      type: object
      required: [login, email]
      properties:
        login:
          type: string
        email:
          type: string
    UserCreateResponse:
      type: object
      required: [login, one_time_password]
      properties:
        login:
          type: string
        one_time_password:
          type: string
    UserUpdate:
      type: object
      properties:
        email:
          type: string
        is_super_admin:
          type: boolean

    # -- Project --
...
    APIKey:
      type: object
      properties:
        id:
          type: integer
          format: int64
        prefix:
          type: string
        label:
          type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
    APIKeyCreate:
      type: object
      required: [label]
      properties:
        label:
          type: string
    ProjectAPIKeyCreate:
      description: |
        Request body for POST /projects/{name}/api-keys. `role` is optional and
        defaults to "maintainer" at the application layer (D-25). Minting a
        project-scoped key with role="viewer" lets CI pipelines publish with
        the correct least-privilege role (D-23 / D-26).
      type: object
      required: [name]
      properties:
        name:
          type: string
        role:
          type: string
          enum: [maintainer, viewer]
    ProjectAPIKeyCreateResponse:
      description: |
        Shown-once response for POST /projects/{name}/api-keys. `secret` is
        returned on this call only — the caller must store it immediately; it
        cannot be retrieved again.
      type: object
      required: [id, name, prefix, secret, created_at]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        prefix:
          type: string
        secret:
          type: string
        created_at:
          type: string
          format: date-time
    APIKeyCreateResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        prefix:
          type: string
        secret:
          type: string
          description: Shown once at creation time only
        label:
          type: string
        created_at:
          type: string
          format: date-time

    # -- S3 Keys --
...
    StatusOK:
      type: object
      properties:
        status:
          type: string

    # -- Scans --
...
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
...
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    cursorParam:
      name: cursor
      in: query
      schema:
        type: string
paths:
...
  /me/api-keys:
    get:
      operationId: listMyAPIKeys
      tags: [api-keys]
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKey'
        default:
          $ref: '#/components/responses/ValidationError'
    post:
      operationId: createAPIKey
      tags: [api-keys]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/APIKeyCreate'
      responses:
        "201":
          description: API key created (secret shown once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKeyCreateResponse'
        default:
          $ref: '#/components/responses/ValidationError'
  /me/api-keys/{id}:
    delete:
      operationId: deleteAPIKey
      tags: [api-keys]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: API key revoked
        default:
          $ref: '#/components/responses/ValidationError'

  # =========================== PROJECT API KEYS ===========================
  /projects/{name}/api-keys:
    get:
      operationId: listProjectAPIKeys
      tags: [api-keys]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of project-owned API keys (omr_p_*)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKey'
        default:
          $ref: '#/components/responses/ValidationError'
    post:
      operationId: createProjectAPIKey
      tags: [api-keys]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectAPIKeyCreate'
      responses:
        "201":
          description: Project API key created (secret shown once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectAPIKeyCreateResponse'
        default:
          $ref: '#/components/responses/ValidationError'
  /projects/{name}/api-keys/{id}:
    delete:
      operationId: deleteProjectAPIKey
      tags: [api-keys]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Project API key revoked
        default:
          $ref: '#/components/responses/ValidationError'

  # ============================== ME S3 KEYS ==============================
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/cursorParam'
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  next_cursor:
                    type: string
                    nullable: true
        default:
          $ref: '#/components/responses/ValidationError'
    post:
      operationId: createUser
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
        default:
          $ref: '#/components/responses/ValidationError'
  /admin/users/{login}:
    get:
      operationId: getUser
      tags: [admin]
      parameters:
        - name: login
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/ValidationError'
    patch:
      operationId: updateUser
      tags: [admin]
      parameters:
        - name: login
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/ValidationError'
    delete:
      operationId: deleteUser
      tags: [admin]
      parameters:
        - name: login
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOK'
        default:
          $ref: '#/components/responses/ValidationError'

  # ============================== ADMIN AUDIT ==============================
...